javascript now working, map not showing
